requirements.txt
autopep8==1.6.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.8
colorama==0.4.4
debugpy==1.5.1
decorator==4.4.2
entrypoints==0.3
future==0.18.2
idna==3.3
imageio==2.13.0
imageio-ffmpeg==0.4.5
ipykernel==6.5.1
ipython==7.30.0
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
matplotlib-inline==0.1.3
moviepy==1.0.3
nest-asyncio==1.5.1
numpy==1.21.4
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
proglog==0.1.9
prompt-toolkit==3.0.23
pycodestyle==2.8.0
Pygments==2.10.0
python-dateutil==2.8.2
pywin32==302
pyzmq==22.3.0
requests==2.26.0
six==1.16.0
toml==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
